pyproject.toml
[tool.poetry]
name = "dio-bank-jesus"
version = "0.1.0"
description = "Desafio de um banco em FastAPI"
authors = ["GhortheBrute <143045110+GhortheBrute@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {version = "*", extras = ["standard"]}
psycopg2-binary = "*"
alembic = "*"
databases = {extras = ["aiosqlite"], version = "^0.9.0"}
pyjwt = "*"
pydantic-settings = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
